refactor: rename app_router to app_health
