+++ app/api/__init__.py
@@ -1,7 +1,7 @@
 """app.api module"""
 from fastapi import APIRouter
 
-from app.api.routes.app_router import router as app_router
+from app.api.routes.app_health import router as app_router
 
 router = APIRouter()
